refactor tests to use FinanceServiceHomeProtocol
