+++ solutions/devsprint-leonardo-santos-3/FinanceAppTests/Modules/Home/HomeViewModelTests.swift
@@ -4,35 +4,35 @@ import XCTest
 
 final class HomeViewModelTests: XCTestCase {
     private let delegateSpy = HomeViewModelDelegateSpy()
-    private let financeSpy = FinanceServiceSpy()
+    private let financeHomeSpy = FinanceHomeServiceSpy()
     private let dispatchQueueSpy = DispatchQueueSpy()
     
     private lazy var sut = HomeViewModel(
-            financeService: financeSpy,
+            financeService: financeHomeSpy,
             dispatchQueue: dispatchQueueSpy
     )
     
     func test_fetchData_givenNilData_shouldNotCallDelegate(){
-        financeSpy.fetchHomeDataCompletionDataToBeReturned = nil
+        financeHomeSpy.fetchHomeDataCompletionDataToBeReturned = nil
         
         sut.fetchData()
         
-        XCTAssertTrue(financeSpy.fetchHomeDataCalled)
-        XCTAssertNil(financeSpy.fetchHomeDataCompletionDataToBeReturned)
-        XCTAssertFalse(dispatchQueueSpy.asyncCalled)
+        XCTAssertTrue(financeHomeSpy.fetchHomeDataCalled)
+        XCTAssertNil(financeHomeSpy.fetchHomeDataCompletionDataToBeReturned)
+        XCTAssertFalse(dispatchQueueSpy.calledDispatchQueueAsync)
         XCTAssertFalse(delegateSpy.didFetchHomeDataCalled)
     }
     
     func test_fetchData_givenValidaData_shouldCallDelegate(){
         sut.delegate = delegateSpy
         
-        financeSpy.fetchHomeDataCompletionDataToBeReturned = HomeData.fixture(balance: 0, savings: 0, spending: 0, activity: [])
+        financeHomeSpy.fetchHomeDataCompletionDataToBeReturned = HomeData.fixture(balance: 0, savings: 0, spending: 0, activity: [])
         
         sut.fetchData()
         
-        XCTAssertTrue(financeSpy.fetchHomeDataCalled)
-        XCTAssertNotNil(financeSpy.fetchHomeDataCompletionDataToBeReturned)
-        XCTAssertTrue(dispatchQueueSpy.asyncCalled)
+        XCTAssertTrue(financeHomeSpy.fetchHomeDataCalled)
+        XCTAssertNotNil(financeHomeSpy.fetchHomeDataCompletionDataToBeReturned)
+        XCTAssertTrue(dispatchQueueSpy.calledDispatchQueueAsync)
         XCTAssertTrue(delegateSpy.didFetchHomeDataCalled)
     }
 }
